전남대 BE_김지환 1단계 - 엔티티 매핑 (#119)
* feat: JPA 의존성 추가 및 application.properties에 관련 설정 추가

* feat: Product를 JPA 엔티티로 변환

* feat: ProductRepository를 Jdbc 기반에서 JPA 기반으로 변경

* refactor: ProductRepository에 의존하는 WishService 및 WishServiceTest 리팩토링

* feat: JPA 적용으로 인한 ProductService 로직 변경

* feat: DataJpaTest를 이용한 ProductRepository 테스트 코드 추가

* feat: Member를 JPA 엔티티로 변환

* feat: MemberRepository를 Jdbc 기반에서 JPA 기반으로 변경

* feat: JPA 적용으로 인한 MemberService 로직 변경

* feat: DataJpaTest를 이용한 MemberRepository 테스트 코드 추가

* refactor: 불필요한 테스트 코드 파일 삭제

* refactor: schema.sql 삭제

* feat: Wishlist를 JPA 엔티티로 변환

* feat: WishlistRepository를 Jdbc 기반에서 JPA 기반으로 변경

* feat: JPA 적용으로 인한 MemberService 로직 변경 및 관련 예외 처리 추가

* refactor: MemberService 로직 변경으로 인한 WishlistController 리팩토링

* feat: JPA의 데이터베이스 스키마 자동 생성을 위한 설정 추가 및 data.sql 사용을 위한 설정 추가

* feat: DataJpaTest를 이용한 WishlistRepository 테스트 코드 추가

+++ src/main/java/gift/member/service/MemberService.java
@@ -10,10 +10,12 @@ import gift.member.entity.Member;
 import gift.member.repository.MemberRepository;
 import gift.member.util.PasswordUtil;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
+@Transactional
 public class MemberService {
     private final MemberRepository memberRepository;
     private final PasswordUtil passwordUtil;
@@ -40,6 +42,7 @@ public class MemberService {
         return MemberResponseDto.from(memberRepository.save(member));
     }
 
+    @Transactional(readOnly = true)
     public List<MemberResponseDto> getMembers() {
         return memberRepository.findAll()
                 .stream()
@@ -47,6 +50,7 @@ public class MemberService {
                 .toList();
     }
 
+    @Transactional(readOnly = true)
     public MemberResponseDto getMemberById(Long id) {
         Member member = memberRepository.findById(id)
                 .orElseThrow(() -> new MemberNotFoundByIdException(id));
@@ -54,6 +58,7 @@ public class MemberService {
         return MemberResponseDto.from(member);
     }
 
+    @Transactional(readOnly = true)
     public Member getMemberByEmail(String email) {
         Member member = memberRepository.findByEmail(email)
                 .orElseThrow(() -> new MemberNotFoundByEmailException(email));
@@ -80,18 +85,20 @@ public class MemberService {
             hashedPassword = passwordUtil.hashPassword(newPassword, salt);
         }
 
-        Member memberToUpdate = new Member(
-                id,
+        member.updateMember(
                 updateRequestDto.email(),
                 salt,
                 hashedPassword,
-                updateRequestDto.role()
-        );
+                updateRequestDto.role());
 
-        return MemberResponseDto.from(memberRepository.update(memberToUpdate));
+        return MemberResponseDto.from(member);
     }
 
     public void deleteMember(Long id) {
+        if(!memberRepository.existsById(id)) {
+            throw new MemberNotFoundByIdException(id);
+        }
+
         memberRepository.deleteById(id);
     }
 }
